only when CA file is copied
could be `if ( ! empty( $options['verify'] ) ) {`

+++ php/commands/core.php
@@ -147,12 +147,16 @@ class Core_Command extends WP_CLI_Command {
 
 		try {
 			$request = Requests::get( $url, $headers, $options );
-			unlink( $options['verify'] );
+			if ( $options['verify'] ) {
+				unlink( $options['verify'] );
+			}
 			return $request;
 		} catch( Requests_Exception $ex ) {
 			// Handle SSL certificate issues gracefully
 			WP_CLI::warning( $ex->getMessage() );
-			unlink( $options['verify'] );
+			if ( $options['verify'] ) {
+				unlink( $options['verify'] );
+			}
 			$options['verify'] = false;
 			try {
 				return Requests::get( $url, $headers, $options );
